fix(worktree): address PR review feedback
- Only send watch_path in DeprioritizeProject for worktree sessions,
  not all sessions (fixes ref-count leak for non-worktree projects)
- Store canonical watchPath from daemon response in sessionState to
  avoid path mismatch on macOS (/private/var vs /var)
- Add isWorktree and watchPath fields to SessionState
- Log I/O errors in find_main_worktree_path instead of silently
  discarding them (distinguish NotFound from permission errors)
- Log DB errors in queue processor worktree lookup instead of
  silent .unwrap_or(None)
- Extract worktree logic from registration.rs into worktree.rs module
  to bring registration.rs under 500-line limit
