+++ src/rust/daemon/grpc/src/services/project_service/mod.rs
@@ -15,6 +15,7 @@ pub(crate) mod lsp_lifecycle;
 mod mutations;
 mod queries;
 mod registration;
+mod worktree;
 
 #[cfg(test)]
 mod tests;
